[ALPS07893511] reviser: fix oob-read
Modify parameter type from u16 to u32

MTK-Commit-Id: 80f8aae6c31d7719e627166d74db35ed692bdfc4

CR-Id: ALPS07893511
Feature: apu_sys_kernel

+++ drivers/misc/mediatek/apusys/reviser/2.0/common/reviser_dbg.c
@@ -63,8 +63,8 @@ static uint32_t g_rvr_debug_op;
 static uint32_t g_rvr_remote_klog;
 static uint32_t g_reviser_vlm_ctx;
 static uint32_t g_reviser_mem_tcm_bank;
-static uint16_t g_reviser_mem_dram_bank;
-static uint16_t g_reviser_mem_dram_ctx;
+static uint32_t g_reviser_mem_dram_bank;
+static uint32_t g_reviser_mem_dram_ctx;
 
 //----------------------------------------------
 // user table dump
